Refactored TextDisplay export
