+++ components/confirmation-modal/confirmation-modal.tsx
@@ -6,7 +6,7 @@ import Button from '../button/button';
 import InstagramButton from '../instagram-button/instagram-button';
 import ModalWrapper from '../modal-wrapper/modal-wrapper';
 import Spinner from '../spinner/spinner';
-import { TextDisplay } from '../text-display/text-display';
+import TextDisplay from '../text-display/text-display';
 
 type InfoModalProps = {
   isVisible: boolean;
