+++ pages/active-auctions/index.tsx
@@ -7,7 +7,7 @@ import AuctionCard from '@/components/auction-card/auction-card';
 import HeadElement from '@/components/head-element/head-element';
 import LinkButton from '@/components/link-button/link-button';
 import Spinner from '@/components/spinner/spinner';
-import { TextDisplay } from '@/components/text-display/text-display';
+import TextDisplay from '@/components/text-display/text-display';
 
 import routes from '../../utilities/routes';
