+++ pages/download/index.tsx
@@ -4,7 +4,7 @@ import { useIntl } from 'react-intl';
 
 import HeadElement from '@/components/head-element/head-element';
 import LinkButton from '@/components/link-button/link-button';
-import { TextDisplay } from '@/components/text-display/text-display';
+import TextDisplay from '@/components/text-display/text-display';
 
 const DownloadPage: FC = (): JSX.Element => {
   const intl = useIntl();
