+++ pages/racing/index.tsx
@@ -6,7 +6,7 @@ import HeadElement from '@/components/head-element/head-element';
 import ImageGrid from '@/components/image-grid/image-grid';
 import ImageModal from '@/components/image-modal/image-modal';
 import Jumbotron from '@/components/jumbotron/jumbotron';
-import { TextDisplay } from '@/components/text-display/text-display';
+import TextDisplay from '@/components/text-display/text-display';
 import { BuildGridImages } from '@/helpers/build-grid-images';
 import { images as img } from '@/public/images/racing';
 import { StaticImage } from '@/types/image-types';
